回退fraction和time容错性修改

+++ rosen/modules/render_service_base/src/drawable/rs_render_node_drawable_adapter.cpp
@@ -173,7 +173,7 @@ void RSRenderNodeDrawableAdapter::DrawRangeImpl(
         return;
     }
 
-    if (end > static_cast<int8_t>(drawCmdList_.size())) {
+    if (static_cast<uint32_t>(end) > drawCmdList_.size()) {
         ROSEN_LOGE("RSRenderNodeDrawableAdapter::DrawRangeImpl, end is invalid");
         return;
     }
@@ -200,7 +200,7 @@ void RSRenderNodeDrawableAdapter::DrawRangeImpl(
 
 void RSRenderNodeDrawableAdapter::DrawImpl(Drawing::Canvas& canvas, const Drawing::Rect& rect, int8_t index) const
 {
-    if (drawCmdList_.empty() || index < 0 || index >= static_cast<int8_t>(drawCmdList_.size())) {
+    if (drawCmdList_.empty() || index < 0 || static_cast<uint32_t>(index) >= drawCmdList_.size()) {
         return;
     }
